Build: Adjust for changes introduced in Atoms 6
In version 6 symbols compiled into `AtomsCore`, AtomsGraph` and
`AtomsUtils` are all in the `Atoms` libraries now. To adjust for this
we use only add those to the linked libraries for versions prior 6.

Additionally Toolchefs moved a `using half` directive into the
appropriate Atoms namespace, but including `AtomsUtils.h` pulls their
shipped `half.h` into the source file which leads to an undefined half
type. We attemt to fix this by instructing the preprocessor to include
the `half.h` found on the given search paths first.

The decision for this is to remove a constraint to add an include directive for half.h in every
source file which might depend on the half data type (via nested includes and
templated definitions provided by Cortex and Gaffer).