CMakeLists.txt: Better handling of cxx11_abi flag
This uses the compiler version instead of the maya version as a
more reliable way to determine if we need to support the legacy
string handling in c++.

This is useful when building for standalone gaffer, and should
also work for maya.

As we believe that ATOMS_ROOT is not used anywhere, we also
removed support for it. If we still need to support it, we'd have
to fix the condition there since ATOMS_MAYA_ROOT would always be
defined.